Refactor generic parts of buffering TestDox printer
